Generate audio scripts at runtime
For some reason, loading certain KeepCoding modules after TV have a chance to nullify all of the AudioScript components that are present in the module. At the moment, there is no way of fixing this, so I removed all of the AudioScripts and AudioSources from the videoplayer, and now the components are added when the module is initialized.

This is not a proper fix, and I do not know if a proper fix is possible. But for the time being, all audio will be controlled by the game's music, due to me not being able to access KeepCoding's _isSFX variable.